Updated Reader, my project imploded.
When running it throws a 50 line exception, so long that I can't google search it. When I try to search it on stack overflow it yields 0 answers. Looks like I won't be able to finish in time. :(
I'll visit CS dojo and see if I can get this patched up.